wip(trash): begin vue all trash (#178)
* wip(trash): begin vue all trash

* feat(menu): add trash link

* wip(trash): correction refacto

* wip(trash): controller

* fix(trash): correction controller

* feat(trash): twig

* fix(innerHTML): remplace by append

* wip(innerHTML): add select multiple

* feat(assets): update querySelector

* wip(multiselect): disable if id is undefined

* fix merge

* fix(js): link enable

* wip(trash): js enable

* feat(trash): support trash admin

* fix php

* fix(ci/cd): add phpstan

* revert(ci/cd): disable phpstan

* fix: delete useless folder

* wip(multiple): twig & js for entity

* fix scss

* fix: correction titre

* feat(multiple): add Btn Restore selection

* feat(multiple): destroy / restore / delete

* fix PHP

+++ apps/src/Controller/Api/ActionsController.php
@@ -2,10 +2,13 @@
 
 namespace Labstag\Controller\Api;
 
+use Exception;
 use Labstag\Annotation\IgnoreSoftDelete;
 use Labstag\Entity\Attachment;
 use Labstag\Lib\ApiControllerLib;
+use Labstag\Service\TrashService;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -14,6 +17,109 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class ActionsController extends ApiControllerLib
 {
+    /**
+     * @Route("/empties", name="api_action_empties", methods={"DELETE"})
+     * @IgnoreSoftDelete
+     *
+     * @return Response
+     */
+    public function empties(Request $request): JsonResponse
+    {
+        $data       = [
+            'action'  => false,
+            'message' => '',
+        ];
+        $tokenValid = $this->apiActionsService->verifToken('empties');
+        if (!$tokenValid) {
+            $data['message'] = 'token incorrect';
+
+            return new JsonResponse($data);
+        }
+
+        $entities = explode(',', $request->request->get('entities'));
+        $error    = [];
+        foreach ($entities as $entity) {
+            $repository = $this->apiActionsService->getRepository($entity);
+            try {
+                $this->deleteEntityByRepository($repository);
+            } catch (Exception $exception) {
+                $error[] = $exception->getMessage();
+            }
+        }
+
+        $data['message'] = $error;
+        if (0 === count($error)) {
+            $data['action'] = true;
+        }
+
+        return new JsonResponse($data);
+    }
+
+    /**
+     * @Route("/emptyall", name="api_action_emptyall", methods={"DELETE"})
+     * @IgnoreSoftDelete
+     *
+     * @return Response
+     */
+    public function emptyall(TrashService $trashService): JsonResponse
+    {
+        $tokenValid = $this->apiActionsService->verifToken('emptyall');
+
+        $data = [
+            'action'  => false,
+            'message' => '',
+        ];
+        if (!$tokenValid) {
+            $data['message'] = 'token incorrect';
+
+            return new JsonResponse($data);
+        }
+
+        $error           = $this->deleteAll($trashService);
+        $data['message'] = $error;
+        if (0 === count($error)) {
+            $data['action'] = true;
+        }
+
+        return new JsonResponse($data);
+    }
+
+    private function deleteAll(TrashService $trashService)
+    {
+        $all   = $trashService->all();
+        $error = [];
+        foreach ($all as $data) {
+            $entity     = $data['name'];
+            $repository = $this->apiActionsService->getRepository($entity);
+            try {
+                $this->deleteEntityByRepository($repository);
+            } catch (Exception $exception) {
+                $error[] = $exception->getMessage();
+            }
+        }
+
+        return $error;
+    }
+
+    private function deleteEntityByRepository($repository)
+    {
+        $all   = $repository->findTrashForAdmin();
+        $files = [];
+        foreach ($all as $entity) {
+            $this->entityManager->remove($entity);
+            if ($entity instanceof Attachment) {
+                $files[] = $entity->getName();
+            }
+        }
+
+        $this->entityManager->flush();
+        foreach ($files as $file) {
+            if ('' != $file && is_file($file)) {
+                unlink($file);
+            }
+        }
+    }
+
     /**
      * @Route("/empty/{entity}", name="api_action_empty", methods={"DELETE"})
      * @IgnoreSoftDelete
@@ -34,24 +140,71 @@ class ActionsController extends ApiControllerLib
             return new JsonResponse($data);
         }
 
-        $data['action'] = true;
-        $all            = $repository->findTrashForAdmin();
-        $files          = [];
-        foreach ($all as $entity) {
-            $this->entityManager->remove($entity);
-            if ($entity instanceof Attachment) {
-                $files[] = $entity->getName();
-            }
+        $error = [];
+        try {
+            $this->deleteEntityByRepository($repository);
+        } catch (Exception $exception) {
+            $error[] = $exception->getMessage();
         }
 
-        $this->entityManager->flush();
-        foreach ($files as $file) {
-            if ('' != $file && is_file($file)) {
-                unlink($file);
-            }
+        $data['error'] = $error;
+        if (0 === count($error)) {
+            $data['action'] = true;
         }
 
         return new JsonResponse($data);
+    }
+
+    /**
+     * @Route("/restories/{entity}", name="api_action_restories", methods={"POST"})
+     * @IgnoreSoftDelete
+     *
+     * @return Response
+     */
+    public function restories(string $entity, Request $request): JsonResponse
+    {
+        $data = [
+            'action'  => false,
+            'message' => '',
+        ];
+
+        $tokenValid = $this->apiActionsService->verifToken('restories');
+        if (!$tokenValid) {
+            $data['message'] = 'token incorrect';
+
+            return new JsonResponse($data);
+        }
+
+        $entities   = explode(',', $request->request->get('entities'));
+        $error      = [];
+        $repository = $this->apiActionsService->getRepository($entity);
+        foreach ($entities as $id) {
+            try {
+                $entity = $repository->find($id);
+                dump($entity);
+                $this->restoreEntity($entity);
+            } catch (Exception $exception) {
+                $error[] = $exception->getMessage();
+            }
+        }
+
+        $data['error'] = $error;
+        if (0 === count($error)) {
+            $data['action'] = true;
+        }
+
+        return new JsonResponse($data);
+    }
+
+    private function restoreEntity($entity)
+    {
+        if (is_null($entity) || is_null($entity->getDeletedAt())) {
+            return;
+        }
+
+        $entity->setDeletedAt(null);
+        $this->entityManager->persist($entity);
+        $this->entityManager->flush();
     }
 
     /**
@@ -82,11 +235,67 @@ class ActionsController extends ApiControllerLib
         }
 
         $data['action'] = true;
-        $entity->setDeletedAt(null);
-        $this->entityManager->persist($entity);
-        $this->entityManager->flush();
+        $this->restoreEntity($entity);
 
         return new JsonResponse($data);
+    }
+
+    /**
+     * @Route("/destroies/{entity}", name="api_action_destroies", methods={"DELETE"})
+     * @IgnoreSoftDelete
+     *
+     * @return Response
+     */
+    public function destroies(string $entity, Request $request): JsonResponse
+    {
+        $data = [
+            'action'  => false,
+            'message' => '',
+        ];
+
+        $tokenValid = $this->apiActionsService->verifToken('destroies');
+        if (!$tokenValid) {
+            $data['message'] = 'token incorrect';
+
+            return new JsonResponse($data);
+        }
+
+        $entities   = explode(',', $request->request->get('entities'));
+        $error      = [];
+        $repository = $this->apiActionsService->getRepository($entity);
+        foreach ($entities as $id) {
+            try {
+                $entity = $repository->find($id);
+                $this->destroyEntity($entity);
+            } catch (Exception $exception) {
+                $error[] = $exception->getMessage();
+            }
+        }
+
+        $data['error'] = $error;
+        if (0 === count($error)) {
+            $data['action'] = true;
+        }
+
+        return new JsonResponse($data);
+    }
+
+    private function destroyEntity($entity)
+    {
+        if (is_null($entity) || is_null($entity->getDeletedAt())) {
+            return;
+        }
+
+        $file = '';
+        $this->entityManager->remove($entity);
+        if ($entity instanceof Attachment) {
+            $file = $entity->getName();
+        }
+
+        $this->entityManager->flush();
+        if ('' != $file && is_file($file)) {
+            unlink($file);
+        }
     }
 
     /**
@@ -117,17 +326,60 @@ class ActionsController extends ApiControllerLib
         }
 
         $data['action'] = true;
-        $this->entityManager->remove($entity);
-        if ($entity instanceof Attachment) {
-            $file = $entity->getName();
-        }
-
-        $this->entityManager->flush();
-        if ('' != $file && is_file($file)) {
-            unlink($file);
-        }
+        $this->destroyEntity($entity);
 
         return new JsonResponse($data);
+    }
+
+    /**
+     * @Route("/deleties/{entity}/", name="api_action_deleties", methods={"DELETE"})
+     *
+     * @return Response
+     */
+    public function deleties(string $entity, Request $request): JsonResponse
+    {
+        $data = [
+            'action'  => false,
+            'message' => '',
+        ];
+
+        $tokenValid = $this->apiActionsService->verifToken('deleties');
+        if (!$tokenValid) {
+            $data['message'] = 'token incorrect';
+
+            return new JsonResponse($data);
+        }
+
+        $entities   = explode(',', $request->request->get('entities'));
+        $error      = [];
+        $repository = $this->apiActionsService->getRepository($entity);
+        foreach ($entities as $id) {
+            try {
+                $entity = $repository->find($id);
+                $this->deleteEntity($entity);
+            } catch (Exception $exception) {
+                $error[] = $exception->getMessage();
+            }
+        }
+
+        $data['error'] = $error;
+        if (0 === count($error)) {
+            $data['action'] = true;
+        }
+
+        unset($entity);
+
+        return new JsonResponse($data);
+    }
+
+    private function deleteEntity($entity)
+    {
+        if (is_null($entity) || !is_null($entity->getDeletedAt())) {
+            return;
+        }
+
+        $this->entityManager->remove($entity);
+        $this->entityManager->flush();
     }
 
     /**
@@ -157,8 +409,7 @@ class ActionsController extends ApiControllerLib
         }
 
         $data['action'] = true;
-        $this->entityManager->remove($entity);
-        $this->entityManager->flush();
+        $this->deleteEntity($entity);
 
         return new JsonResponse($data);
     }
